set the demopayment inder the suspence and resolve the vercel deployment issus
