+++ mnt/data/transactions.csv
@@ -1,2 +1,12 @@
 _id,transaction_id,user_id,amount,payment_method,card_type,merchant_id,previous_txn_count_24h,avg_amount_30d,country,city,location,is_fraud,status,timestamp,__v
 68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
+68d1b30793807825a8d59ba3,txn_68d1b2ee93807825a8d59b9f,undefined,1800,MOBILEBANKING,NAGAD-Nagad,lily68cc567288bd2,0,1800,US,Mountain View,"37.4056,-122.0775",,success,2025-09-22T20:35:19.230Z,0
